Fix conflicts with existing genrule rule
Previously, the `genalex` and `genhappy` macros would create genrules
with the name based on the source file. This can cause conflicts when
the same source file occurs multiple times. This may happen if it
corresponds to different outputs each time.

The macros `genalex` and `genhappy` are called for each `alex` or
`happy` source found in a subdirectory of every entry in the
`hs-source-dirs` Cabal section. Consider the following directory
structure.

```
foo
└── bar
    └── baz.x
```

And the following `hs-source-dirs` section:

```
hs-source-dirs:
  foo
  foo/bar
```

This will cause `foo/bar/baz.x` to occur twice. Once with the output
`gen-srcs-pkg/bar/baz.hs` and once with the output
`gen-srcs-pkg/baz.hs`. If the `genrule` name is based on the input, then
this will cause a conflict. However, if the `genrule` name is bazed on
the output it will not.

Since Hazel does not refer to the `genrule` name anywhere it is safe to
change it.